fix: collect classes with cfg from models
