+++ src/events/models.py
@@ -1,27 +1,15 @@
 from django.db import models
 from django.utils.safestring import mark_safe
 
-
-class EventsCfg:
-    """
-    Настройки для модели Events.
-    """
-
-    MAX_LENGTH_DEFAULT = 255
-    EVENTS_MAX_LENGTH = 5000
-    EVENTS_HELP_MSG_NAME = "Введите название События"
-    EVENTS_HELP_MSG_TXT = (
-        f"Напишите текст события до {EVENTS_MAX_LENGTH} символов"
-    )
-    EVENTS_HELP_MSG_IMG = "Добавьте ссылки на изображения с событиями"
+from core.constants import MAX_LENGTH_DEFAULT, EventsCfg, MAX_LENGTH
 
 
 class Events(models.Model):
     id = models.AutoField(primary_key=True)
     name = models.CharField(
-        max_length=EventsCfg.MAX_LENGTH_DEFAULT,
+        max_length=MAX_LENGTH_DEFAULT,
         verbose_name="Событие",
-        help_text=EventsCfg.EVENTS_HELP_MSG_TXT,
+        help_text=EventsCfg.EVENTS_HELP_MSG_NAME,
     )
     image = models.ManyToManyField(
         "EventsImageURL",
@@ -35,7 +23,7 @@ class Events(models.Model):
     )
     text = models.TextField(
         verbose_name="Содержание события",
-        max_length=EventsCfg.EVENTS_MAX_LENGTH,
+        max_length=MAX_LENGTH,
         help_text=EventsCfg.EVENTS_HELP_MSG_TXT,
     )
 
@@ -56,7 +44,7 @@ class EventsImageURL(models.Model):
         verbose_name="События",
     )
     image = models.URLField(
-        max_length=EventsCfg.MAX_LENGTH_DEFAULT,
+        max_length=MAX_LENGTH_DEFAULT,
         unique=True,
         verbose_name="Ссылка на изображение события",
         help_text="Укажите URL-адрес изображения о событии",
