+++ src/sports/models.py
@@ -1,24 +1,12 @@
 from django.db import models
 from django.utils.safestring import mark_safe
 
-
-class SportsCfg:
-    """
-    Настройки для модели Sports.
-    """
-
-    MAX_LENGTH_DEFAULT = 255
-    SPORTS_MAX_LENGTH = 5000
-    SPORTS_HELP_MSG_NAME = "Введите название Вида спорта"
-    SPORTS_HELP_MSG_TXT = (
-        f"Напишите описание вида спорта до {SPORTS_MAX_LENGTH} символов"
-    )
-    EVENTS_HELP_MSG_IMG = "Добавьте ссылки на изображения с видами спорта"
+from core.constants import MAX_LENGTH_DEFAULT, SportsCfg, MAX_LENGTH
 
 
 class Sports(models.Model):
     name = models.CharField(
-        max_length=SportsCfg.MAX_LENGTH_DEFAULT,
+        max_length=MAX_LENGTH_DEFAULT,
         verbose_name="Вид спорта",
         help_text=SportsCfg.SPORTS_HELP_MSG_NAME,
     )
@@ -30,7 +18,7 @@ class Sports(models.Model):
     )
     text = models.TextField(
         verbose_name="Описание вида спорта",
-        max_length=SportsCfg.SPORTS_MAX_LENGTH,
+        max_length=MAX_LENGTH,
         help_text=SportsCfg.SPORTS_HELP_MSG_TXT,
     )
 
@@ -51,7 +39,7 @@ class SportsImageURL(models.Model):
         verbose_name="Виды спорта",
     )
     image = models.URLField(
-        max_length=SportsCfg.MAX_LENGTH_DEFAULT,
+        max_length=MAX_LENGTH_DEFAULT,
         unique=True,
         verbose_name="Ссылка на изображение с видом спорта",
         help_text="Укажите URL-адрес изображения с видом спорта",
